requirements.txt
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu124
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
accelerate==1.6.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antialiased-cnns==0.3
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git
cmake==4.0.0
contourpy==1.3.2
cycler==0.12.1
diffusers==0.33.1
docker-pycreds==0.4.0
dominate==2.9.1
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
gdown==4.4.0
gitdb==4.0.12
GitPython==3.1.44
gradio==3.43.1
gradio_client==0.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
lit==18.1.8
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.35.0
networkx==3.4.2
numpy==1.26.4
open_clip_torch==2.32.0
opencv-python==4.6.0.66
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peft==0.15.2
pillow==10.4.0
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.11.1
semantic-version==2.10.0
sentry-sdk==2.26.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
streamlit==1.44.1
streamlit-keyup==0.2.0
sympy==1.13.1
tenacity==9.1.2
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torchdata==0.11.0
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.2
vision-aided-loss==0.1.0
wandb==0.19.10
watchdog==6.0.0
wcwidth==0.2.13
websockets==11.0.3
xformers==0.0.29.post3
zipp==3.21.0
